chore: 폴더구조 적용및 ESLint + husky 설정
